+++ code/pyorg/scripts/groel/bi_2nd_analysis.py
@@ -1,17 +1,17 @@
 """
 
-    Performs Univariate 2nd order analysis to a SetListTomograms object by tomograms
+    Performs Bivariate 2nd order analysis to a SetListTomograms object by tomograms
 
-    Input:  - A STAR file with a set of ListTomoParticles pickles (SetListTomograms object input)
-            - Analysis settings
+    Input:  - A STAR file with a ListTomoParticles pickle (SetListTomograms object input) used as reference
+            - A STAR file with a set of ListTomoParticles pickles (SetListTomograms object input) to analyze
 
     Output: - Plots by tomograms:
                 + Number of particles by list
                 + Particles densities by list
                 + Plots by list:
-                    * Univariate 2nd order analysis against random simulation
+                    * Bivariate 2nd order analysis against random simulation
             - Global plots:
-                + Global univariate 2nd order analysis against random simulation by list
+                + Global Bivariate 2nd order analysis against random simulation by list
 
 """
 
@@ -25,9 +25,11 @@ import sys
 import time
 from pyorg import pexceptions, sub, disperse_io, surf
 from pyorg.globals import unpickle_obj, clean_dir
-from pyorg.surf.model import ModelCSRV, gen_tlist
+from pyorg.surf.model import ModelCSRV
 from pyorg.surf.utils import list_tomoparticles_pvalues
+import matplotlib.pyplot as plt
 from matplotlib import pyplot as plt, rcParams
+
 try:
     import pickle as pickle
 except ImportError:
@@ -47,45 +49,51 @@ rcParams['ytick.labelsize'] = 14
 # PARAMETERS
 ########################################################################################
 
-ROOT_PATH = '/fs/pool/pool-ruben/antonio/groel' # '/fs/pool/pool-ruben/Jonathan/PhD/clustering_analysis_antonio'
+ROOT_PATH = '/fs/pool/pool-ruben/antonio/groel' # '/fs/pool/pool-plitzko/Saikat/luminal_particle_organization/lattice_break_clustering'
 
 # Input STAR file
-in_star = ROOT_PATH + '/ltomos/tomo12006_all/all_ltomos.star' # '/ltomos/test_reduced/test_reduced_ltomos.star'
-in_wspace = None # ROOT_PATH + '/tests/test_reduced/test_reduced_20_400_10_10_sim_20_wspace.pkl' # (Insert a path to recover a pickled workspace instead of doing a new computation)
+in_star_ref = ROOT_PATH + '/ltomos/tomo12006_all/all_ltomos.star' # '/ltomos/v1_nobc_proj/v1_nobc_proj_ltomos.star'
+in_ref_short_key = '4' # '0'
+in_star = ROOT_PATH + '/ltomos/tomo12006_all/all_ltomos.star' # '/ltomos/v1_nobc_proj/v1_nobc_proj_ltomos.star'  #
+in_wspace = None # ROOT_PATH + '/data/tutorials/synth_sumb/org/uni_2nd/test_1_ref_0_proj_8_300_10_sim_10_wspace.pkl'  # (Insert a path to recover a pickled workspace instead of doing a new computation)
 
 # Output directory
-out_dir = ROOT_PATH + '/ana/tomo12006' # '/tests/test_reduced/'
-out_stem = 'uni_2nd_8_200_5_rdf_10_sim_10_over0.05'
+out_dir = ROOT_PATH + '/ana/tomo12006' # '/org/bi/bi_2nd/'
+out_stem = 'bi_2nd_ref_4_8_300_3_sim_20_rdf_10_switch_over0.05' # 'v1_nobc_proj_sph_10_600_10'  # 'uni_sph_4_60_5'
 
 # Analysis variables
-ana_res = 1.408 # nm/voxel
-ana_rg = np.arange(8, 300, 5) # in nm
-ana_shell_thick = 15 # None # 5
-ana_rdf = True
-ana_conv_iter = None # 100
-ana_max_iter = None # 100000
+ana_res = 1.368  # nm/voxel
+ana_rg = np.arange(8, 300, 3)  # in nm
+ana_shell_thick = 10 # 5 # None
+ana_rdf = True # True
 ana_fmm = False
-ana_npr = 10 # None means Auto
 
 # P-value computation settings
 # Simulation model (currently only CSRV)
-p_nsims = 10 # 50
-p_per = 5 # %
+p_nsims = 20
+p_per = 5  # %
+
+# Multiprocessing settings
+mp_npr = 10 # None means Auto
 
 ##### Advanced settings
+
+ana_conv_iter = None # 100
+ana_max_iter = None # 100000
 
 ana_border = True
 
 # Particle surface
-p_vtp = None # ROOT_PATH + '/in/bref_16_14.08_cent_dec.vtp'
-p_cinter = 0.05 # Maximum fraction of overlapping between particles
+p_vtp = None
+p_cinter = 0.05
+p_switch = True  # False
 
 # Figure saving options
-fig_fmt = '.png' # if None they are displayed instead
+fig_fmt = '.png'  # if None they showed instead
 
 # Plotting options
-pt_sim_v = True
 pt_cmap = plt.get_cmap('gist_rainbow')
+
 
 ########################################################################################
 # MAIN ROUTINE
@@ -102,6 +110,7 @@ def compute_ic(per, sims):
     ic_high = np.percentile(sims, 100 - per, axis=0, interpolation='linear')
     return ic_low, ic_med, ic_high
 
+
 # Computes pvalue from a matrix of simulations (n_arrays, array_samples)
 def compute_pvals(exp_med, sims):
     n_sims = float(sims.shape[0])
@@ -111,6 +120,7 @@ def compute_pvals(exp_med, sims):
         p_vals[i] = float((exp >= sim_slice).sum()) / n_sims
     return p_vals
 
+
 # Units conversion
 ana_rg_v = ana_rg / ana_res
 ana_shell_thick_v = None
@@ -119,13 +129,15 @@ if ana_shell_thick is not None:
 
 ########## Print initial message
 
-print('Univariate second order analysis for a ListTomoParticles by tomograms.')
+print('Bivariate 1:N second order analysis for a ListTomoParticles by tomograms.')
 print('\tAuthor: ' + __author__)
 print('\tDate: ' + time.strftime("%c") + '\n')
 print('Options:')
 print('\tOutput directory: ' + str(out_dir))
 print('\tOuput stem: ' + str(out_stem))
-print('\tInput STAR file: ' + str(in_star))
+print('\tInput reference STAR file: ' + str(in_star_ref))
+print('\tInput reference short key: ' + str(in_ref_short_key))
+print('\tInput analysis STAR file: ' + str(in_star))
 if in_wspace is not None:
     print('\tLoad workspace from: ' + in_wspace)
 print('\tOrganization analysis settings: ')
@@ -137,9 +149,7 @@ else:
     print('\t\t-Shell neighborhood with thickness: ' + str(ana_shell_thick) + ' nm')
     print('\t\t-Shell neighborhood with thickness: ' + str(ana_shell_thick_v) + ' voxels')
     if ana_rdf:
-        print('\t\t-Computing Radial Distribution Function.')
-    else:
-        print('\t\t-Computing Function-O.')
+        print('\t\t-RDF activated.')
 if in_wspace is None:
     if (ana_conv_iter is not None) and (ana_max_iter is not None):
         print('\t\t-Convergence number of samples for stochastic volume estimations: ' + str(ana_conv_iter))
@@ -149,10 +159,10 @@ if in_wspace is None:
             print('\t\t-FMM for distance metric computation.')
         else:
             print('\t\t-DT for distance transform computation.')
-    if ana_npr is None:
+    if mp_npr is None:
         print('\t\t-Number of processors: Auto')
     else:
-        print('\t\t-Number of processors: ' + str(ana_npr))
+        print('\t\t-Number of processors: ' + str(mp_npr))
 print('\tP-Value computation setting:')
 print('\t\t-Percentile: ' + str(p_per) + ' %')
 print('\t\t-Number of instances for simulations: ' + str(p_nsims))
@@ -160,14 +170,14 @@ print('\tSurface settings:')
 if p_vtp is not None:
     print('\t\t-Particle surface: ' + p_vtp)
 print('\t\t-Maximum fraction of overlapping surface for simulated particles:' + str(p_cinter))
+if p_switch:
+    print('\t\t-Random simulations using the reference pattern.')
 if fig_fmt is not None:
     print('\tStoring figures:')
     print('\t\t-Format: ' + str(fig_fmt))
 else:
     print('\tPlotting settings: ')
 print('\t\t-Colormap: ' + str(pt_cmap))
-if pt_sim_v:
-    print('\t\t-Verbose simulation activated!')
 print('')
 
 ######### Process
@@ -181,35 +191,52 @@ if os.path.exists(out_stem_dir):
     clean_dir(out_stem_dir)
 else:
     os.makedirs(out_stem_dir)
-out_tomos_dir = out_stem_dir + '/tomos'
-os.makedirs(out_tomos_dir)
+
+print('\tLoading input data...')
+star, star_ref = sub.Star(), sub.Star()
+try:
+    star.load(in_star)
+    star_ref.load(in_star_ref)
+except pexceptions.PySegInputError as e:
+    print('ERROR: input STAR file could not be loaded because of "' + e.get_message() + '"')
+    print('Terminated. (' + time.strftime("%c") + ')')
+    sys.exit(-1)
+ref_list = None
+for row in range(star_ref.get_nrows()):
+    ltomos_pkl = star_ref.get_element('_psPickleFile', row)
+    fname_pkl = os.path.split(ltomos_pkl)[1]
+    try:
+        idx = fname_pkl.index('_')
+    except ValueError:
+        continue
+    if fname_pkl[:idx] == in_ref_short_key:
+        ref_list = unpickle_obj(ltomos_pkl)
+if ref_list is None:
+    print('ERROR: reference ListTomoParticles with short key ' + in_ref_short_key + ' was not found!')
+    print('Terminated. (' + time.strftime("%c") + ')')
+    sys.exit(-1)
 
 if in_wspace is None:
 
-    print('\tLoading input data...')
-    star = sub.Star()
-    try:
-        star.load(in_star)
-    except pexceptions.PySegInputError as e:
-        print('ERROR: input STAR file could not be loaded because of "' + e.get_message() + '"')
-        print('Terminated. (' + time.strftime("%c") + ')')
-        sys.exit(-1)
-    set_lists, lists_dic_rows = surf.SetListTomoParticles(), dict()
+    set_lists = surf.SetListTomoParticles()
+    part_vtps = dict()
     for row in range(star.get_nrows()):
         ltomos_pkl = star.get_element('_psPickleFile', row)
         ltomos = unpickle_obj(ltomos_pkl)
         set_lists.add_list_tomos(ltomos, ltomos_pkl)
-        fkey = os.path.split(ltomos_pkl)[1]
-        short_key_idx = fkey.index('_')
-        short_key = fkey[:short_key_idx]
-        lists_dic_rows[short_key] = row
-    if p_vtp is not None:
         try:
-            part_vtp = disperse_io.load_poly(p_vtp)
+            if p_vtp is not None:
+                part_vtp = disperse_io.load_poly(p_vtp)
+            else:
+                part_vtp = disperse_io.load_poly(star.get_element('_suSurfaceVtpSim', row))
         except pexceptions.PySegInputError as e:
             print('ERROR: reference particle surface file could not be loaded because of "' + e.get_message() + '"')
             print('Terminated. (' + time.strftime("%c") + ')')
             sys.exit(-1)
+        lkey = os.path.split(ltomos_pkl)[1]
+        short_key_idx = lkey.index('_')
+        short_key = lkey[:short_key_idx]
+        part_vtps[short_key] = part_vtp
 
     print('\tBuilding the dictionaries...')
     lists_count, tomos_count = 0, 0
@@ -237,6 +264,11 @@ if in_wspace is None:
         llist_tomos_dic = llist.get_tomos()
         for tkey, ltomo in zip(iter(llist_tomos_dic.keys()), iter(llist_tomos_dic.values())):
             try:
+                ref_tomo = ref_list.get_tomo_by_key(tkey)
+            except KeyError:
+                print('\t\t\t\t\t-WARNING: tomogram ' + tkey + ' discarded because it is not in reference list!')
+                continue
+            try:
                 tomos_np[tkey]
             except KeyError:
                 tomos_hash[tkey] = tomos_count
@@ -252,45 +284,23 @@ if in_wspace is None:
             tomos_exp[tkey][lkey], tomos_sim[tkey][lkey] = list(), list()
 
     print('\tLIST COMPUTING LOOP:')
-    sim_obj_set = surf.SetListSimulations()
-    for lkey in lists_hash.values():
+    for lkey in list(lists_hash.values()):
+
+        if (in_star_ref in in_star) and (lkey == in_ref_short_key):
+            print('\t\t-Processing and reference list of particles are the same, skipping...')
+            continue
 
         llist = lists_dic[lkey]
-        sim_obj_list = surf.ListSimulations()
         print('\t\t-Processing list: ' + lkey)
-        hold_row = lists_dic_rows[lkey]
-        if p_vtp is None:
-            try:
-                p_vtp = star.get_element('_suSurfaceVtpSim', hold_row)
-            except KeyError:
-                print('ERROR: _suSurfaceVtpSim is required in the input STAR file "' + in_star + '"')
-                print('Terminated. (' + time.strftime("%c") + ')')
-                sys.exit(-1)
-        part_vtp = disperse_io.load_poly(p_vtp)
-
+        part_vtp = part_vtps[lkey]
         print('\t\t\t+Tomograms computing loop:')
-        for tkey in tomos_hash.keys():
+        for tkey in list(tomos_hash.keys()):
 
-            print('\t\t\t\t*Processing tomogram: ' + tkey) # os.path.split(tkey)[1]
-            try:
-                ltomo = llist.get_tomo_by_key(tkey)
-            except KeyError:
-                print('\t\t\t\t\t-Tomogram with key ' + tkey + ' is not in the list ' + lkey + ' , continuing...')
+            print('\t\t\t\t*Processing tomogram: ' + os.path.split(tkey)[1])
+            ltomo, ref_tomo = llist.get_tomo_by_key(tkey), ref_list.get_tomo_by_key(tkey)
+            if ref_tomo.get_num_particles() <= 0:
+                print('\t\t\t\t\t\t+No reference particles found, skipping...')
                 continue
-            if ltomo.get_num_particles() <= 0:
-                print('\t\t\t\t\t-WARNING: no particles to process, continuing...')
-                continue
-
-            # if p_nsims > 0:
-            #     tkey_short = os.path.splitext(os.path.split(tkey)[1])[0]
-            #     hold_dir = out_tomos_dir + '/' + tkey.replace('/', '_')
-            #     if not os.path.exists(hold_dir):
-            #         os.makedirs(hold_dir)
-                # out_sim_model = hold_dir + '/model_' + str(lkey) + '.vtp'
-                # print '\t\t\t\t+Storing a simulation model in: ' + out_sim_model
-                # model_sim = ModelCSRV(ltomo.get_voi(), part_vtp)
-                # sim_tomos = gen_tlist(1, ltomo.get_num_particles(), model_sim, mode_emb='center', npr=1)
-                # disperse_io.save_vtp(sim_tomos.get_tomo_list()[0].append_particles_vtp(mode='surface'), out_sim_model)
 
             print('\t\t\t\t\t-Computing the number of particles...')
             hold_np = ltomo.get_num_particles()
@@ -302,30 +312,31 @@ if in_wspace is None:
             tomos_den[tkey].append(hold_den)
             lists_den[lkey][tomos_hash[tkey]] = hold_den
 
-            print('\t\t\t\t\t-Computing univariate second order metrics...')
-            hold_arr = ltomo.compute_uni_2nd_order(ana_rg_v, thick=ana_shell_thick_v, border=ana_border,
-                                                   conv_iter=ana_conv_iter, max_iter=ana_max_iter, fmm=ana_fmm,
-                                                   rdf=ana_rdf, npr=ana_npr)
+            print('\t\t\t\t\t-Computing bivariate second order metrics...')
+            hold_arr = ref_tomo.compute_bi_2nd_order(ltomo, ana_rg_v, thick=ana_shell_thick_v, border=ana_border,
+                                                     conv_iter=ana_conv_iter, max_iter=ana_max_iter, fmm=ana_fmm,
+                                                     rdf=ana_rdf, npr=mp_npr)
             if hold_arr is not None:
                 tomos_exp[tkey][lkey].append(hold_arr)
-                for i in range(ltomo.get_num_particles()):
+                for i in range(ref_tomo.get_num_particles()):
                     lists_exp[lkey].append(hold_arr)
 
-            print('\t\t\t\t\t-Simulating univariate second order metrics...')
+            print('\t\t\t\t\t-Simulating bivariate second order metrics...')
             model_csrv = ModelCSRV(check_inter=p_cinter)
-            hold_arr = ltomo.simulate_uni_2nd_order(p_nsims, model_csrv, part_vtp, 'center', ana_rg_v, border=ana_border,
-                                                    thick=ana_shell_thick_v, fmm=ana_fmm,
-                                                    conv_iter=ana_conv_iter, max_iter=ana_max_iter, rdf=ana_rdf,
-                                                    npr=ana_npr)
+            # hold_arr = ltomo.simulate_uni_2nd_order(p_nsims, model_csrv, part_vtp, 'center', ana_rg_v,
+            #                                         border=ana_border,
+            #                                         thick=ana_shell_thick_v, fmm=ana_fmm,
+            #                                         conv_iter=ana_conv_iter, max_iter=ana_max_iter, rdf=ana_rdf,
+            #                                         npr=mp_npr)
+            hold_arr = ref_tomo.simulate_bi_2nd_order(ltomo, p_nsims, model_csrv, part_vtp, 'center',
+                                                      ana_rg_v, thick=ana_shell_thick_v, border=ana_border,
+                                                      conv_iter=ana_conv_iter, max_iter=ana_max_iter, rdf=ana_rdf,
+                                                      fmm=ana_fmm, npr=mp_npr, switched=p_switch)
             if hold_arr is not None:
-                sim_obj = surf.Simulation(ana_rg, ana_shell_thick, np.array(hold_arr))
-                sim_obj_list.insert_simulation(tkey, sim_obj)
                 for arr in hold_arr:
                     tomos_sim[tkey][lkey].append(arr)
-                    for i in range(ltomo.get_num_particles()):
+                    for i in range(ref_tomo.get_num_particles()):
                         lists_sim[lkey].append(arr)
-
-        sim_obj_set.insert_list_simulations(lkey, sim_obj_list)
 
     out_wspace = out_dir + '/' + out_stem + '_wspace.pkl'
     print('\tPickling computation workspace in: ' + out_wspace)
@@ -337,16 +348,10 @@ if in_wspace is None:
         pickle.dump(wspace, fl)
         fl.close()
 
-    out_lsims = out_dir + '/' + out_stem + '_lsims.pkl'
-    print('\tPickling the list of simulations results: ' + out_lsims)
-    with open(out_lsims, "wb") as fl:
-        print(type(sim_obj_set))
-        pickle.dump(sim_obj_set, fl)
-        fl.close()
-
 else:
+
     print('\tLoading the workspace: ' + in_wspace)
-    with open(in_wspace, 'rb') as pkl:
+    with open(in_wspace, 'r') as pkl:
         wspace = pickle.load(pkl)
     lists_count, tomos_count = wspace[0], wspace[1]
     lists_hash, tomos_hash = wspace[2], wspace[3]
@@ -360,12 +365,20 @@ print('\tPrinting tomograms hash: ')
 for tkey, val in zip(iter(tomos_hash.keys()), iter(tomos_hash.values())):
     print('\t\t-[' + tkey + '] -> [' + str(val) + ']')
 
+print('\tPrinting number of samples and volume for the reference by tomogram: ')
+for tomo in ref_list.get_tomo_list():
+    tkey = tomo.get_tomo_fname()
+    print('\t\t-' + tkey + ': ' + str(tomo.get_num_particles()) + ' np, ' + str(tomo.compute_voi_volume()) + ' nm**3')
+
 # Getting the lists colormap
 n_lists = len(list(lists_hash.keys()))
 for i, lkey in zip(iter(lists_hash.keys()), iter(lists_hash.values())):
-    lists_color[lkey] = pt_cmap(1.*i/n_lists)
+    lists_color[lkey] = pt_cmap(1. * i / n_lists)
 
 print('\tTOMOGRAMS PLOTTING LOOP: ')
+
+out_tomos_dir = out_stem_dir + '/tomos'
+os.makedirs(out_tomos_dir)
 
 print('\t\t-Plotting the number of particles...')
 for tkey, ltomo in zip(iter(tomos_np.keys()), iter(tomos_np.values())):
@@ -383,6 +396,7 @@ for tkey, ltomo in zip(iter(tomos_np.keys()), iter(tomos_np.values())):
     if fig_fmt is None:
         plt.show(block=True)
     else:
+        # hold_dir = out_tomos_dir + '/' + tkey_short
         hold_dir = out_tomos_dir + '/' + tkey.replace('/', '_')
         if not os.path.exists(hold_dir):
             os.makedirs(hold_dir)
@@ -405,6 +419,7 @@ for tkey, ltomo in zip(iter(tomos_den.keys()), iter(tomos_den.values())):
     if fig_fmt is None:
         plt.show(block=True)
     else:
+        # hold_dir = out_tomos_dir + '/' + tkey_short
         hold_dir = out_tomos_dir + '/' + tkey.replace('/', '_')
         if not os.path.exists(hold_dir):
             os.makedirs(hold_dir)
@@ -412,6 +427,8 @@ for tkey, ltomo in zip(iter(tomos_den.keys()), iter(tomos_den.values())):
     plt.close()
 
 print('\t\t-Plotting 2nd order metric...')
+low_pvals, high_pvals = list(), list()
+mask_rg = (ana_rg >= 7) & (ana_rg <= 100)
 for tkey, ltomo in zip(iter(tomos_exp.keys()), iter(tomos_exp.values())):
     p_values = dict()
     tkey_short = os.path.splitext(os.path.split(tkey)[1])[0]
@@ -426,49 +443,41 @@ for tkey, ltomo in zip(iter(tomos_exp.keys()), iter(tomos_exp.values())):
             print('\t\t\t+WARNING: no data for tomogram ' + tkey + ' and list ' + lkey)
             continue
         plt.figure()
-        plt.title('Univariate 2nd order for ' + tkey_short + ' and ' + lkey)
+        plt.title('Bivariate 2nd order for ' + tkey_short + ' and ' + lkey)
         if ana_shell_thick is None:
-            plt.ylabel('Function L')
+            plt.ylabel('Ripley\'s L')
         else:
             if ana_rdf:
                 plt.ylabel('Radial Distribution Function')
             else:
-                plt.ylabel('Function O')
-        plt.xlabel('Scale [nm]')
-        if lkey == '0':
-            plt.plot(ana_rg, arr, color=lists_color[lkey], label='CONTROL')
-        elif lkey == '1':
-            plt.plot(ana_rg, arr, color=lists_color[lkey], label='RAPA')
-        else:
-            plt.plot(ana_rg, arr, color=lists_color[lkey], label=lkey)
-
+                plt.ylabel('Ripley\'s O')
+        plt.xlabel('Scale (nm)')
+        plt.plot(ana_rg, arr, color=lists_color[lkey], label=lkey)
         plt.legend(loc=4)
-        try:
-            # plt.plot(ana_rg, ic_low, 'k--')
-            plt.plot(ana_rg, ic_med, 'k', label='SIMULATED')
-            # plt.plot(ana_rg, ic_high, 'k--')
-        except NameError:
-            pass
-        plt.fill_between(ana_rg, ic_low, ic_high, alpha=0.5, color='gray', edgecolor='w')
-        # plt.ylim(-9, 8)
-        # plt.grid(True)
+        plt.plot(ana_rg, ic_low, 'k--')
+        plt.plot(ana_rg, ic_med, 'k')
+        plt.plot(ana_rg, ic_high, 'k--')
         plt.tight_layout()
         if fig_fmt is None:
             plt.show(block=True)
         else:
+            # hold_dir = out_tomos_dir + '/' + tkey_short
             hold_dir = out_tomos_dir + '/' + tkey.replace('/', '_')
             if not os.path.exists(hold_dir):
                 os.makedirs(hold_dir)
-            plt.savefig(hold_dir + '/uni_' + lkey + '.png')
+            plt.savefig(hold_dir + '/bi_' + lkey + '.png')
         plt.close()
 
     plt.figure()
     plt.title('Clustering p-value for ' + tkey_short)
     plt.ylabel('p-value')
-    plt.xlabel('Scale [nm]')
+    plt.xlabel('Scale (nm)')
     try:
         for lkey, vals in zip(iter(p_values.keys()), iter(p_values.values())):
             plt.plot(ana_rg, vals, color=lists_color[lkey], linestyle='-', label=lkey)
+            hold_vals = vals[mask_rg]
+            high_pvals.append(hold_vals.max())
+            low_pvals.append(1 - vals.min())
     except IndexError:
         print('\t\t\t+WARNING: no p-values for list: ' + lkey)
     plt.legend(loc=4)
@@ -476,59 +485,31 @@ for tkey, ltomo in zip(iter(tomos_exp.keys()), iter(tomos_exp.values())):
     if fig_fmt is None:
         plt.show(block=True)
     else:
+        # hold_dir = out_tomos_dir + '/' + tkey_short
         hold_dir = out_tomos_dir + '/' + tkey.replace('/', '_')
         if not os.path.exists(hold_dir):
             os.makedirs(hold_dir)
         plt.savefig(hold_dir + '/pvals.png')
     plt.close()
 
-print('\t\t-Plotting 2nd order metric (all simulations)...')
+print('\t\t-Plotting p-values box (tomo[s_low, s_high]= (min_pval, max_pval)):')
 plt.figure()
-# plt.title('Univariate 2nd order for all simulations')
-hold_sim_vals = dict()
-max_nsims = p_nsims
-if max_nsims > 5:
-    max_nsims = 5
-for tkey, ltomo in zip(iter(tomos_exp.keys()), iter(tomos_exp.values())):
-    p_values = dict()
-    tkey_short = os.path.splitext(os.path.split(tkey)[1])[0]
-    for lkey, arr in zip(iter(tomos_exp[tkey].keys()), iter(tomos_exp[tkey].values())):
-        try:
-            hold_sims = np.asarray(tomos_sim[tkey][lkey], dtype=np.float32)
-            for hold_sim in hold_sims[:max_nsims]:
-                plt.plot(ana_rg, hold_sim, color=lists_color[lkey], alpha=.3)
-                try:
-                    hold_sim_vals[lkey].append(hold_sim)
-                except KeyError:
-                    hold_sim_vals[lkey] = list()
-                    hold_sim_vals[lkey].append(hold_sim)
-        except IndexError:
-            print('\t\t\t+WARNING: no data for tomogram ' + tkey + ' and list ' + lkey)
-            continue
-        if ana_shell_thick is None:
-            plt.ylabel('Function L')
-        else:
-            if ana_rdf:
-                plt.ylabel('Radial Distribution Function')
-            else:
-                plt.ylabel('Function O')
-                plt.ticklabel_format(style='sci', axis='y', scilimits=(0, 0))
-for lkey, vals in zip(iter(hold_sim_vals.keys()), iter(hold_sim_vals.values())):
-    vals_mat = np.asarray(vals, dtype=np.float32)
-    lbl = lkey
-    if lkey == '0':
-        lbl = 'CONTROL'
-    elif lkey == '1':
-        lbl = 'RAPA'
-    plt.plot(ana_rg, np.median(vals_mat, axis=0), color=lists_color[lkey], label=lbl, linewidth=3.0, linestyle='--')
-plt.xlabel('Scale [nm]')
-# plt.grid(True)
-plt.legend(loc=4)
+if ana_shell_thick is None:
+    plt.title('Ripley\'s L p-values box-plot')
+    plt.ylabel('Ripley\'s L (p-values)')
+else:
+    if ana_rdf:
+        plt.title('RDF p-values box-plot')
+        plt.ylabel('RDF (p-values)')
+    else:
+        plt.title('Ripley\'s O p-values box-plot')
+        plt.ylabel('Ripley\'s O (p-values)')
+plt.boxplot([low_pvals, high_pvals], labels=['Low', 'High'])
 plt.tight_layout()
 if fig_fmt is None:
     plt.show(block=True)
 else:
-    plt.savefig(out_tomos_dir + '/uni_sims.png')
+    plt.savefig(out_tomos_dir + '/Box_pvals.png')
 plt.close()
 
 print('\tLISTS PLOTTING LOOP: ')
@@ -553,114 +534,50 @@ for lkey, tlist in zip(iter(lists_np.keys()), iter(lists_np.values())):
         hold_dir = out_lists_dir + '/' + lkey
         if not os.path.exists(hold_dir):
             os.makedirs(hold_dir)
-        plt.savefig(hold_dir + '/np_lists.png')
+        plt.savefig(hold_dir + '/np.png')
     plt.close()
 
 print('\t\t-Plotting densities...')
-plt.figure()
-# plt.title('Densities for ' + lkey)
-plt.ylabel('Density [#/$\mu m^3$]')
-# plt.xlabel('Tomograms')
-count, ticks = 1, dict()
-ana_res_3_i = 1. / float((1e-3 * ana_res)**3)
-plt.ticklabel_format(style='sci', axis='y', scilimits=(0,0))
 for lkey, tlist in zip(iter(lists_den.keys()), iter(lists_den.values())):
-    arr = tlist[np.where(tlist > 0)[0]]
-    den_low, den_med, den_high = np.percentile(arr, p_per), np.percentile(arr, 50), np.percentile(arr, 100 - p_per)
-    den_low, den_med, den_high = den_low * ana_res_3_i, den_med * ana_res_3_i, den_high * ana_res_3_i
-    plt.bar(.5 * count, den_med, width=.40, color=lists_color[lkey], edgecolor='k', linewidth=2)
-    plt.errorbar(.5 * count, den_med, yerr=np.asarray([[den_med - den_low, den_high - den_med], ]).reshape(2, 1),
-                 ecolor='k', elinewidth=4, capthick=4, capsize=8)
-    if lkey == '0':
-        ticks['CONTROL'] = .5 * count
-    elif lkey == '1':
-        ticks['RAPA'] = .5 * count
+    plt.figure()
+    plt.title('Densities for ' + lkey)
+    plt.ylabel('Density (np/vol)')
+    plt.xlabel('Tomograms')
+    for i, den in enumerate(tlist):
+        plt.bar(i, den, width=0.75)
+    plt.xticks(BAR_WIDTH + np.arange(n_tomos), np.arange(n_tomos))
+    plt.tight_layout()
+    if fig_fmt is None:
+        plt.show(block=True)
     else:
-        ticks[lkey] = .5 * count
-    count += 1
-plt.xticks(list(ticks.values()), list(ticks.keys()))
-plt.tight_layout()
-if fig_fmt is None:
-    plt.show(block=True)
-else:
-    plt.savefig(out_lists_dir + '/den_lists.png')
-plt.close()
+        hold_dir = out_lists_dir + '/' + lkey
+        if not os.path.exists(hold_dir):
+            os.makedirs(hold_dir)
+        plt.savefig(hold_dir + '/den.png')
+    plt.close()
 
-print('\t\t-Plotting gathered 2nd order metric...')
-sims, do_pval = list(), True
-exps_low, exps_med, exps_high = list(), list(), list()
-plt.figure()
-# plt.title('Univariate 2nd order for ' + lkey)
-if ana_shell_thick is None:
-    plt.ylabel('Function L')
-else:
-    if ana_rdf:
-        plt.ylabel('Radial Distribution Function')
-        plt.plot((0, ana_rg[-1]), (1, 1), linestyle='--', linewidth=1.0, color='k')
-    else:
-        plt.ylabel('Function O')
-        plt.ticklabel_format(style='sci', axis='y', scilimits=(0, 0))
-plt.xlabel('Scale [nm]')
 for lkey, tlist in zip(iter(lists_exp.keys()), iter(lists_exp.values())):
-    # Getting experimental IC
-    exp_low, exp_med, exp_high = compute_ic(p_per, np.asarray(tlist))
-    # plt.plot(ana_rg, exp_low, color=lists_color[lkey], linestyle='--', label=lkey)
-    plt.plot(ana_rg, exp_med, color=lists_color[lkey], linestyle='-', linewidth=2.0, label=lkey)
-    # plt.plot(ana_rg, exp_high, color=lists_color[lkey], linestyle='--', label=lkey)
-    plt.fill_between(ana_rg, exp_low, exp_high, alpha=0.3, color=lists_color[lkey], edgecolor='w')
-    sims += lists_sim[lkey]
-    exps_low.append(exp_low)
-    exps_med.append(exp_med)
-    exps_high.append(exp_high)
-# Getting simulations IC
-hold_sim = np.asarray(sims)
-hold_sim[np.isnan(hold_sim) | np.isinf(hold_sim)] = 0
-ic_low, ic_med, ic_high = compute_ic(p_per, np.asarray(hold_sim))
-try:
-    # plt.plot(ana_rg, ic_low, 'k--')
-    plt.plot(ana_rg, ic_med, 'k', linewidth=2.0)
-    # plt.plot(ana_rg, ic_high, 'k--')
-except (NameError, ValueError):
-    do_pval = False
-plt.fill_between(ana_rg, ic_low, ic_high, alpha=0.5, color='gray', edgecolor='w')
-plt.tight_layout()
-plt.xlim((0, ana_rg[-1]))
-# plt.ylim((-9, 8))
-# extraticks = [5,]
-# plt.xticks(list(plt.xticks()[0]) + extraticks)
-if fig_fmt is None:
-    plt.show(block=True)
-else:
-    # plt.show(block=True)
-    plt.savefig(out_lists_dir + '/uni_lists_gathered_' + str(lkey) + '.png', dpi=600)
-plt.close()
-
-print('\t\t-Plotting 2nd order metric...')
-sims, do_pval = list(), True
-exps_low, exps_med, exps_high = list(), list(), list()
-for lkey, tlist in zip(iter(lists_exp.keys()), iter(lists_exp.values())):
+    if len(tlist) <= 0:
+        continue
     plt.figure()
     # plt.title('Univariate 2nd order for ' + lkey)
     if ana_shell_thick is None:
-        plt.ylabel('Function L')
+        plt.ylabel('Ripley\'s L')
     else:
         if ana_rdf:
             plt.ylabel('Radial Distribution Function')
-            plt.plot((0, ana_rg[-1]), (1, 1), linestyle='--', linewidth=1.0, color='k')
         else:
-            plt.ylabel('Function O')
-            plt.ticklabel_format(style='sci', axis='y', scilimits=(0, 0))
+            plt.ylabel('Ripley\'s O')
     plt.xlabel('Scale [nm]')
     # Getting experimental IC
-    exp_low, exp_med, exp_high = compute_ic(p_per, np.asarray(tlist))
+    hold_arr = np.asarray(tlist)
+    hold_arr[np.isnan(hold_arr) | np.isinf(hold_arr)] = 0
+    exp_low, exp_med, exp_high = compute_ic(p_per, hold_arr)
     # plt.plot(ana_rg, exp_low, color=lists_color[lkey], linestyle='--', label=lkey)
-    plt.plot(ana_rg, exp_med, color=lists_color[lkey], linestyle='-', linewidth=2.0, label=lkey)
+    plt.plot(ana_rg, exp_med, color=lists_color[lkey], linestyle='-', label=lkey, linewidth=2.0)
     # plt.plot(ana_rg, exp_high, color=lists_color[lkey], linestyle='--', label=lkey)
     plt.fill_between(ana_rg, exp_low, exp_high, alpha=0.3, color=lists_color[lkey], edgecolor='w')
-    sims += lists_sim[lkey]
-    exps_low.append(exp_low)
-    exps_med.append(exp_med)
-    exps_high.append(exp_high)
+    # sims += lists_sim[lkey]
     # Getting simulations IC
     hold_sim = np.asarray(lists_sim[lkey])
     hold_sim[np.isnan(hold_sim) | np.isinf(hold_sim)] = 0
@@ -674,41 +591,67 @@ for lkey, tlist in zip(iter(lists_exp.keys()), iter(lists_exp.values())):
     plt.fill_between(ana_rg, ic_low, ic_high, alpha=0.5, color='gray', edgecolor='w')
     plt.tight_layout()
     plt.xlim((0, ana_rg[-1]))
-    # plt.ylim((-9, 8))
-    # extraticks = [5,]
-    # plt.xticks(list(plt.xticks()[0]) + extraticks)
     if fig_fmt is None:
         plt.show(block=True)
     else:
         # plt.show(block=True)
-        plt.savefig(out_lists_dir + '/uni_lists_' + str(lkey) + '.png', dpi=600)
+        plt.savefig(out_lists_dir + '/bi_lists_' + str(lkey) + '.png', dpi=600)
     plt.close()
-sims = np.asarray(hold_sim)
-out_fspace = out_dir + '/' + out_stem + '_fspace.pkl'
-print('\tPickling figue in: ' + out_fspace)
-fspace = (ana_rg,
-          exps_low, exps_med, exps_high,
-          ic_low, ic_med, ic_high)
-with open(out_fspace, "wb") as fl:
-    pickle.dump(fspace, fl)
-    fl.close()
 
-if do_pval:
-    print('\t\t-Plotting clustering p-value...')
+print('\t\t-Plotting 2nd order metric...')
+sims = list()
+for lkey, tlist in zip(iter(lists_exp.keys()), iter(lists_exp.values())):
+    if len(tlist) <= 0:
+        continue
     plt.figure()
-    plt.title('Clustering p-value')
-    plt.ylabel('p-value')
-    plt.xlabel('Scale [nm]')
-    for lkey, tlist in zip(iter(lists_exp.keys()), iter(lists_exp.values())):
-        exp_med = compute_ic(p_per, np.asarray(tlist))[1]
-        p_values = compute_pvals(exp_med, sims)
-        plt.plot(ana_rg, p_values, color=lists_color[lkey], linestyle='-', label=lkey)
-    plt.legend(loc=4)
+    plt.title('Bivariate 2nd order for ' + lkey)
+    if ana_shell_thick is None:
+        plt.ylabel('Ripley\'s L')
+    else:
+        if ana_rdf:
+            plt.ylabel('Radial Distribution Function')
+        else:
+            plt.ylabel('Ripley\'s O')
+    plt.xlabel('Scale (nm)')
+    # Getting experimental IC
+    exp_low, exp_med, exp_high = compute_ic(p_per, np.asarray(tlist))
+    plt.plot(ana_rg, exp_low, color=lists_color[lkey], linestyle='--', label=lkey)
+    plt.plot(ana_rg, exp_med, color=lists_color[lkey], linestyle='-', label=lkey)
+    plt.plot(ana_rg, exp_high, color=lists_color[lkey], linestyle='--', label=lkey)
+    # Getting simulations IC
+    hold_sim = lists_sim[lkey]
+    ic_low, ic_med, ic_high = compute_ic(p_per, np.asarray(hold_sim))
+    plt.plot(ana_rg, ic_low, 'k--')
+    plt.plot(ana_rg, ic_med, 'k')
+    plt.plot(ana_rg, ic_high, 'k--')
     plt.tight_layout()
     if fig_fmt is None:
         plt.show(block=True)
     else:
-        plt.savefig(out_lists_dir + '/pvals_lists.png')
+        plt.savefig(out_lists_dir + '/bi_list_' + lkey + '.png')
     plt.close()
+sims = np.asarray(hold_sim)
+
+print('\t\t-Plotting clustering p-value...')
+plt.figure()
+plt.title('Co-localization significance')
+plt.ylabel('p-value')
+plt.xlabel('Scale [nm]')
+for lkey, tlist in zip(iter(lists_exp.keys()), iter(lists_exp.values())):
+    if len(tlist) <= 0:
+        continue
+    exp_med = compute_ic(p_per, np.asarray(tlist))[1]
+    p_values = compute_pvals(exp_med, sims)
+    plt.plot(ana_rg, p_values, color=lists_color[lkey], linestyle='-', label=lkey)
+plt.legend(loc=4)
+plt.tight_layout()
+plt.xlim((0, ana_rg[-1]))
+if fig_fmt is None:
+    plt.show(block=True)
+else:
+    plt.savefig(out_lists_dir + '/pvals_lists.png')
+plt.close()
 
 print('Successfully terminated. (' + time.strftime("%c") + ')')
+
+