+++ code/pyorg/scripts/ribo_test/uni_2nd_analysis.py
@@ -39,9 +39,9 @@ __author__ = 'Antonio Martinez-Sanchez'
 
 BAR_WIDTH = .35
 
-rcParams['axes.labelsize'] = 22 # 14
-rcParams['xtick.labelsize'] = 22 # 14
-rcParams['ytick.labelsize'] = 22 # 14
+rcParams['axes.labelsize'] = 14 # 22
+rcParams['xtick.labelsize'] = 14 # 22
+rcParams['ytick.labelsize'] = 14 # 22
 
 ########################################################################################
 # PARAMETERS
@@ -50,26 +50,26 @@ rcParams['ytick.labelsize'] = 22 # 14
 ROOT_PATH = '/fs/pool/pool-engel/antonio/ribo'
 
 # Input STAR file
-in_star = ROOT_PATH + '/ltomos_v2/test_no_pid_p3_proj_swapxy/all_L_proj_ltomos.star' # '/ltomos_v2/all_no_pid_p3_proj_swapxy/all_L_proj_ltomos.star' # '/ltomos_v2/all_no_pid_p3/all_L_ltomos.star'
-in_wspace = None # ROOT_PATH + '/tests/uni_dsa_test_p3_proj_swapxy/test_8_300_3_sim_1_wspace.pkl' # (Insert a path to recover a pickled workspace instead of doing a new computation)
+in_star = ROOT_PATH + '/ltomos_v2/all_no_pid_p3_proj_swapxy/all_L_proj_ltomos.star' # '/ltomos_v2/test_no_pid_p3_proj_swapxy/all_L_proj_ltomos.star' # '/ltomos_v2/all_no_pid_p3/all_L_ltomos.star'
+in_wspace = ROOT_PATH + '/ana_2nd/uni_dsa_test_p3_proj_swapxy_over0/all_8_200_3_sim_10_dt_wspace.pkl' # (Insert a path to recover a pickled workspace instead of doing a new computation)
 
 # Output directory
-out_dir = ROOT_PATH + '/tests/uni_dsa_shell_all_p3/' # '/ana_2nd/uni_dsa_test_p3_proj_swapxy_over0/' # '/tests/uni_dsa_shell_all_p3/'
-out_stem = 'time_test_8_50_3_rdf_10_fmm_sim_1'
+out_dir = ROOT_PATH + '/ana_2nd/uni_dsa_test_p3_proj_swapxy_over0/' # '/tests/uni_dsa_shell_all_p3/' # '/tests/uni_dsa_shell_all_p3/'
+out_stem = 'all_8_200_3_sim_10_dt_v2' # 'time_test_8_50_3_rdf_10_fmm_sim_1'
 
 # Analysis variables
 ana_res = 2.096 # nm/voxel
-ana_rg = np.arange(8, 50, 3) # in nm
-ana_shell_thick = 10 # None # 5
+ana_rg = np.arange(8, 200, 3) # in nm
+ana_shell_thick = None # 10 # 5
 ana_rdf = False # False
 ana_conv_iter = None # 100
 ana_max_iter = None # 100000
-ana_fmm = True # False
-ana_npr = 1 # 10 # None means Auto
+ana_fmm = False
+ana_npr = 10 # None means Auto
 
 # P-value computation settings
 # Simulation model (currently only CSRV)
-p_nsims = 1 # 5 # 50
+p_nsims = 10 # 5 # 50
 p_per = 5 # %
 
 ##### Advanced settings
@@ -156,8 +156,10 @@ if in_wspace is None:
 print('\tP-Value computation setting:')
 print('\t\t-Percentile: ' + str(p_per) + ' %')
 print('\t\t-Number of instances for simulations: ' + str(p_nsims))
+print('\tSurface settings:')
 if p_vtp is not None:
     print('\t\t-Particle surface: ' + p_vtp)
+print('\t\t-Maximum fraction of overlapping surface for simulated particles:' + str(p_cinter))
 if fig_fmt is not None:
     print('\tStoring figures:')
     print('\t\t-Format: ' + str(fig_fmt))
@@ -610,6 +612,8 @@ for lkey, tlist in zip(iter(lists_exp.keys()), iter(lists_exp.values())):
     exps_low.append(exp_low)
     exps_med.append(exp_med)
     exps_high.append(exp_high)
+    # sim_med = compute_ic(p_per, np.asarray(lists_sim[lkey]))[1]
+    # plt.plot(ana_rg, sim_med, color=lists_color[lkey], linestyle='--', linewidth=2.0)
 # Getting simulations IC
 hold_sim = np.asarray(sims)
 hold_sim[np.isnan(hold_sim) | np.isinf(hold_sim)] = 0